chore(ci/release): validate release ref not eq branch (#177)
This prevents common mistake that the ref to release from is the branch
name to which the version bump is made. This is because the workflow
first bumps the version, and then builds/publishes from the release-ref.
If the two are equal, it leads to building the version that is just
bumped.

Signed-off-by: Alex Gronskiy <agronskiy@nvidia.com>