Llane/home pg style edits (#215)
This is a very basic migration of the readme content + adding a minimal
toctree to the home index page so that the sphinx site produces a
sidebar. The sidebar will mature and break out in the future into
sections such as About, Get Started, etc.

We will also add more sections/cards to this page after all other basic
edits have been checked in, so it won't be a direct copy of the README,
instead it will become a proper docs site home page.

---------

Signed-off-by: Lawrence Lane <llane@nvidia.com>
Signed-off-by: L.B. <llane@nvidia.com>
Co-authored-by: jgerh <163925524+jgerh@users.noreply.github.com>
Signed-off-by: Anna Warno <awarno@nvidia.com>