feat(adapter): add optional update for progress tracking on timer (#423)
Inference can be slow depending on the model performance, the number of
GPUs the model is running on, the context length or tokens to generate
or parallelism. For slow inference times, an interval based progress
tracking may be a frustrating user experience to wait 30mins or longer
for the first progress update.

This PR adds an optional parameter to ProgressTracking adapter to update
progress on a timed interval, not enabled by default.
ProgressTrackingInterceptor tracks the last status that was posted, and
skips sending a request / writes to file if there has not been a change.
The timed update is run as a asyncio task that is stopped on
`post_eval_hook`

---------

Signed-off-by: Kim Ngo <6362111+findkim@users.noreply.github.com>