docs(api-reference): NeMo Evaluator Core Python API Reference (#437)
What does this MR entails:

1. Adds Python API Reference for NeMo Evaluator Core (NEC)
2. Moves CLI Reference to References 
3. Change of `autodoc2` to `autodoc` in sphinx. This is to facilitate
using `autodoc_pydantic` which is an essential extension for rendering
pydantic classes. The downside of a former "`autodoc`" is that it
analyzes code through importing (as opposed to static analysis of
`autodoc2`)
4. Configuration options were removed from `Libraries` -> `Core` ->
`Interceptors`. Users are asked to refer to the API
5. API structure is semi-automatic: All classes are rendered
automatically but I decided to give submodules structure myself.
6.` __all__` variables were fixed
7. Some docstrings were updated (especially related to API dataclasses)
8. Removed unused code in NEC - `base.py`, `run.py`, `registry.py`

What is not included in this MR and is going to be included in
follow-ups:

1. NeMo Evaluator Launcher API Reference
2. Better Python API resolution for Core

---------

Signed-off-by: Tomasz Grzegorzek <tgrzegorzek@nvidia.com>
Signed-off-by: prokotg <19536019+prokotg@users.noreply.github.com>
Co-authored-by: Marta Stepniewska-Dziubinska <marta-sd@users.noreply.github.com>